refactor: 주문메뉴 개수 자료형 변경 - int > long

+++ src/main/java/christmas/domain/discount/WeekdayDiscount.java
@@ -33,7 +33,7 @@ public class WeekdayDiscount implements Discount {
     }
 
     private Money calculateProfit() {
-        int desertCount = ordersMenuCount.getDessertCount();
+        long desertCount = ordersMenuCount.getDessertCount();
         return Money.YEAR_AMOUNT.multiply(desertCount).negative();
     }
 }
